+++ Watchdog_2040.c
@@ -12,6 +12,7 @@
 #include "core/data_structures/string_list.h"
 #include "core/hardware_drivers/haptics.h"
 #include "core/hardware_drivers/joystick.h"
+#include "core/hardware_drivers/sdcard.h"
 #include "core/hardware_drivers/ssd1306.h"
 #include "core/tools/menus.h"
 #include "core/utils/path.h"
@@ -32,7 +33,10 @@ void sys_setup() {
   sys_paths_manager_init();
   // wait_for_user_input();
   first_boot_file = path_init(FIRST_BOOT_FILE);
+  bool should_end_loading_screen = true;
   if (is_first_boot()) {
+    end_loading_screen();
+    should_end_loading_screen = false;
     start_malloc_mascot_tutorial();
     sys_paths_manager_load();
     sys_paths_manager_make();
@@ -44,7 +48,8 @@ void sys_setup() {
   sys_paths_manager_load();
   msg_manager_init(malloc_memories_inst->ulmp_addr);
   home_page_init();
-  end_loading_screen();
+  if (should_end_loading_screen)
+    end_loading_screen();
 }
 
 void sys_mainloop() {
@@ -53,17 +58,19 @@ void sys_mainloop() {
   bool ledvalue = false;
   uint8_t screen_up_seconds = 10;
   uint32_t screen_up_start;
+  check_pheripherals();
   while (true) {
     joystick_update(drivers->joystick);
-    check_pheripherals();
     process_system_state();
+    update_conversations();
     if (joystick_get_direction(drivers->joystick) != C || first_run) {
       first_run = false;
       haptic_short_pulse();
       screen_up_start = to_us_since_boot(get_absolute_time()) / 1000000;
       while (true) {
-        check_pheripherals();
+        // check_pheripherals();
         process_system_state();
+        update_conversations();
         display_home_page();
         joystick_update(drivers->joystick);
         if (joystick_get_direction(drivers->joystick) == E) {