+++ apps/notes/notes.h
@@ -1,0 +1,6 @@
+#ifndef NOTES_H
+#define NOTES_H
+
+void enter_notes_submenu();
+
+#endif