+++ apps/system_submenus/system_submenus.h
@@ -1,7 +1,10 @@
 #ifndef SYSTEM_SUBMENUS_H
 #define SYSTEM_SUBMENUS_H
 
-void display_system_info();
+#include <stdbool.h>
+
+void display_system_info(bool serial_output);
+void display_system_info_wrapped();
 void reset_system();
 void display_battery_status();
 void display_joystick_check();