+++ core/components/hw_manager.c
@@ -4,12 +4,11 @@
 #include <stdint.h>
 #include <stdlib.h>
 
-#include "pico/multicore.h"
-
 #include "core/boot/bootup.h"
 #include "core/data_structures/string_list.h"
 #include "core/hardware_drivers/battery.h"
 #include "core/hardware_drivers/config.h"
+#include "core/hardware_drivers/core1.h"
 #include "core/hardware_drivers/ens160.h"
 #include "core/hardware_drivers/haptics.h"
 #include "core/hardware_drivers/joystick.h"
@@ -17,7 +16,6 @@
 #include "core/hardware_drivers/sdcard.h"
 #include "core/hardware_drivers/ssd1306.h"
 #include "core/hardware_drivers/sx1278.h"
-#include "core/utils/utils.h"
 #include "hardware/adc.h"
 #include "hardware/clocks.h"
 
@@ -34,16 +32,16 @@ hw_drivers *hardware_drivers_init() {
       SSD1306_HEIGHT,
       SSD1306_ADDR);
   drivers = hw_man;
-  multicore_launch_core1(display_bootup_screen);
+  core1_launch(display_bootup_screen);
   hw_man->air_quality_sensor = ens160_init(ENS160_SDA,
       ENS160_SCK,
       ENS160_I2C_PORT,
       ENS160_BAUDRATE,
       ENS160_ADDR);
   if (hw_man->air_quality_sensor->is_working)
-    multicore_fifo_push_blocking(ENS160_OK);
+    core1_push_instruction(ENS160_OK);
   else
-    multicore_fifo_push_blocking(ENS160_ERR);
+    core1_push_instruction(ENS160_ERR);
   hw_man->lora_module = sx1278_init(SX1278_MOSI,
       SX1278_MISO,
       SX1278_SCK,
@@ -54,18 +52,18 @@ hw_drivers *hardware_drivers_init() {
       SX1278_TX_POWER,
       NULL);
   if (hw_man->lora_module->is_working)
-    multicore_fifo_push_blocking(SX1278_OK);
+    core1_push_instruction(SX1278_OK);
   else
-    multicore_fifo_push_blocking(SX1278_ERR);
+    core1_push_instruction(SX1278_ERR);
   hw_man->battery = battery_init(ADC_MAX_VALUE,
       BATTERY_MIN_VOLTAGE,
       BATTERY_MAX_VOLTAGE,
       BATTERY_PIN,
       ADC_CHANNEL);
   if (hw_man->battery->is_working)
-    multicore_fifo_push_blocking(BATTERY_OK);
+    core1_push_instruction(BATTERY_OK);
   else
-    multicore_fifo_push_blocking(BATTERY_ERR);
+    core1_push_instruction(BATTERY_ERR);
   hw_man->joystick = joystick_init(JOYSTICK_X_PIN,
       JOYSTICK_Y_PIN,
       JOYSTICK_X_CHANNEL,
@@ -74,31 +72,27 @@ hw_drivers *hardware_drivers_init() {
       JOYSTICK_SENSITIVITY,
       -90);
   if (hw_man->joystick->is_working)
-    multicore_fifo_push_blocking(JOYSTICK_OK);
+    core1_push_instruction(JOYSTICK_OK);
   else
-    multicore_fifo_push_blocking(JOYSTICK_ERR);
+    core1_push_instruction(JOYSTICK_ERR);
   hw_man->sd_card = sdcard_init();
   sdcard_mount(hw_man->sd_card);
   if (hw_man->sd_card->is_working)
-    multicore_fifo_push_blocking(SDCARD_OK);
+    core1_push_instruction(SDCARD_OK);
   else
-    multicore_fifo_push_blocking(SDCARD_ERR);
+    core1_push_instruction(SDCARD_ERR);
   hw_man->rtc = rtc_time_init(2025, 5, 9, 4, 20, 37, 00);
-  multicore_fifo_push_blocking(RTC_OK);
-  multicore_fifo_push_blocking(CHECKS_END);
+  core1_push_instruction(RTC_OK);
+  core1_push_instruction(CHECKS_END);
   return hw_man;
 }
 
 void end_loading_screen() {
-  wait_for_core1();
-  multicore_reset_core1();
+  core1_await();
+  core1_reset();
   haptics_init(HAPTICS_MOTOR_PIN);
+  core1_listens_for_instructions();
   ssd1306_clear(drivers->oled_screen);
-}
-
-void wait_for_core1() {
-  while (multicore_fifo_pop_blocking() != CORE_1_OP_DONE)
-    ;
 }
 
 uint32_t get_total_heap(void) {