+++ core/hardware_drivers/sdcard.h
@@ -38,6 +38,7 @@ bool sdcard_rename(sdcard *sd, path *src, path *dest);
 bool sdcard_touch_file(sdcard *sd, path *file);
 bool sdcard_mkdir(sdcard *sd, path *dir);
 bool sdcard_rmdir(sdcard *sd, path *dir);
+bool sdcard_rmtree(sdcard *sd, path *dir);
 bool sdcard_path_is_dir(sdcard *sd, path *path);
 // void sdcard_format(sdcard *sd);